+++ src/main/java/com/imsweb/naaccrxml/sas/SasFieldInfo.java
@@ -3,6 +3,9 @@
  */
 package com.imsweb.naaccrxml.sas;
 
+import java.util.Arrays;
+import java.util.List;
+
 /**
  * Small DTO to wrap basic field information.
  * <br/><br/>
@@ -19,6 +22,7 @@ public class SasFieldInfo {
     private final Integer _num;
     private final String _name;
     private final Integer _start;
+    private final List<String> _contains;
 
     public SasFieldInfo(String naaccrId, String truncatedNaaccrId, String parentTag, Integer length, Integer num, String name, Integer start) {
         _naaccrId = naaccrId;
@@ -28,6 +32,18 @@ public class SasFieldInfo {
         _num = num;
         _name = name;
         _start = start;
+        _contains = null;
+    }
+
+    public SasFieldInfo(String naaccrId, Integer num, String name, String parentTag, Integer length, String contains) {
+        _naaccrId = naaccrId;
+        _truncatedNaaccrId = null;
+        _parentTag = parentTag;
+        _length = length;
+        _num = num;
+        _name = name;
+        _start = null;
+        _contains = Arrays.asList(contains.split(","));
     }
 
     public String getTruncatedNaaccrId() {
@@ -56,5 +72,9 @@ public class SasFieldInfo {
 
     public Integer getStart() {
         return _start;
+    }
+
+    public List<String> getContains() {
+        return _contains;
     }
 }