+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 **Version 8.9**
 
 - Added new utility method "xmlToXml" in NaaccrXmlUtils class to allow a file to be re-created while applying some logic to each patient it contains.
-- Re-added support for providing a user-defined dictionary when processing a flat file (Standalone program only).
 - Improve validation of start columns for N18 dictionaries and prior.
 - Fixed validation of user-defined dictionaries not detecting grouped items.
+- Re-added support for providing a user-defined dictionary when processing a flat file (Standalone program only).
+- Updated embedded JRE to Java 17.0.2 (this only affects the standalone program).
 
 **Version 8.8**